edit profile form, removing profile pic, validation in patch user endpoint

+++ api/src/Controller/UserController.php
@@ -20,8 +20,8 @@ class UserController extends BaseController
         $body = $this->request->getContent();
 
         if (!isset($body["username"],
-            $body["first_name"],
-            $body["last_name"],
+            $body["firstName"],
+            $body["lastName"],
             $body["email"],
             $body["password"])) {
             return new JsonResponse(["message" => "Missing fields in request body"], 400);
@@ -30,8 +30,8 @@ class UserController extends BaseController
         $username = trim($body["username"]);
         $email = trim($body["email"]);
         $password = trim($body["password"]);
-        $firstName = trim($body["first_name"]);
-        $lastName = trim($body["last_name"]);
+        $firstName = trim($body["firstName"]);
+        $lastName = trim($body["lastName"]);
 
         if (!$this->validateEmail($email) ||
             !$this->validateName($firstName) ||
@@ -123,19 +123,42 @@ class UserController extends BaseController
 
         $data = $this->request->getContent() ?? [];
 
-        if(isset($data["oldPassword"], $data["newPassword"])){
+        $error = null;
+        if (isset($data["oldPassword"], $data["newPassword"])) {
             $oldHashedPassword = hash_hmac('sha256', $data["oldPassword"], $_ENV["PASSWORD_HASH_KEY"]);
-            if($oldHashedPassword !== $user->getPassword()){
-                return new JsonResponse(["message" => "Invalid password"], 400);
+            if ($oldHashedPassword !== $user->getPassword()) {
+                $error = "Invalid password";
+            } else if (!$this->validatePassword($data["newPassword"])) {
+                $error = "Invalid new password";
+            } else {
+                $user->setPassword(hash_hmac('sha256', $data["newPassword"], $_ENV["PASSWORD_HASH_KEY"]));
             }
-            if(!$this->validatePassword($data["newPassword"])){
-                return new JsonResponse(["message" => "Invalid new password"], 400);
+        } else if (isset($data["username"])) {
+            if (!empty($repo->findBy(["username" => $data["username"]]))) {
+                $error = "The user with this username already exists";
+            } else if (!$this->validateUsername($data["username"])) {
+                $error = "Invalid username";
             }
-            $user->setPassword(hash_hmac('sha256', $data["newPassword"], $_ENV["PASSWORD_HASH_KEY"]));
+        } else if (isset($data["email"])) {
+            if (!empty($repo->findBy(["email" => $data["email"]]))) {
+                $error = "The user with this email already exists";
+            } else if (!$this->validateEmail($data["email"])) {
+                $error = "Invalid email";
+            }
+        } else if (isset($data["firstName"]) && !$this->validateName($data["firstName"])) {
+            $error = "Invalid first name";
+        } else if (isset($data["lastName"]) && !$this->validateName($data["lastName"])) {
+            $error = "Invalid last name";
+        }
+
+        if ($error !== null) {
+            return new JsonResponse([
+                "message" => $error
+            ], 400);
         }
 
         if (isset($data["profilePic"]) && $data["profilePic"] === "REMOVE") {
-            if(is_file($_SERVER['DOCUMENT_ROOT'] . "/" . $user->getProfilePic())){
+            if (is_file($_SERVER['DOCUMENT_ROOT'] . "/" . $user->getProfilePic())) {
                 unlink($_SERVER['DOCUMENT_ROOT'] . "/" . $user->getProfilePic());
             }
             $user->setProfilePic(null);
